Type fixes

+++ Race/src/main/java/model/Vehicle.java
@@ -11,5 +11,5 @@ public abstract class Vehicle {
         return speed;
     }
 
-    public abstract float getTime(int length);
+    public abstract float getTime(float length);
 }
