+++ Race/src/main/java/model/air/Broom.java
@@ -6,7 +6,7 @@ public class Broom extends AirVehicle{
     }
 
     @Override
-    public int getDistanceReducer(int length) {
+    public float getDistanceReducer(float length) {
         return length / 1000;
     }
 }
